+++ vscode/CHANGELOG.md
@@ -8,6 +8,10 @@ file. Format inspired by https://keepachangelog.com/en/1.0.0/
 ### Added
 - tag completion within notes
 
+### Changed
+- Show search results container when search completes. Only works if there's
+  at least one search result :(
+
 ## [0.0.11] - 2020-05-06
 ### Changed
 - created note is saved in same directory is currently open note